Refactor data-table table module split
