+++ packages/data-table/src/lib/database.test.ts
@@ -6,7 +6,8 @@ import { column } from './column.ts'
 import { createDatabase, Database } from './database.ts'
 import { query } from './query.ts'
 import { DataTableAdapterError, DataTableQueryError, DataTableValidationError } from './errors.ts'
-import { belongsTo, table, hasMany, hasManyThrough, hasOne, timestamps } from './table.ts'
+import { table, timestamps } from './table.ts'
+import { belongsTo, hasMany, hasManyThrough, hasOne } from './table-relations.ts'
 import { eq } from './operators.ts'
 import { sql } from './sql.ts'
 import type { SqliteTestAdapterOptions, SqliteTestSeed } from '../../test/sqlite-test-database.ts'
