+++ packages/data-table/src/lib/database/relations.ts
@@ -4,8 +4,10 @@ import type { Predicate } from '../operators.ts'
 import { and, eq, inList, or } from '../operators.ts'
 import type { Query } from '../query.ts'
 import { query as createQuery } from '../query.ts'
-import type { AnyRelation, AnyTable, Relation } from '../table.ts'
-import { getCompositeKey, getTableName, getTablePrimaryKey } from '../table.ts'
+import type { AnyRelation, Relation } from '../table-relations.ts'
+import type { AnyTable } from '../table.ts'
+import { getTableName, getTablePrimaryKey } from '../table.ts'
+import { getCompositeKey } from '../table-keys.ts'
 
 import type { QueryExecutionContext } from './execution-context.ts'
 import { loadRowsWithRelationsForQuery } from './query-execution.ts'
