+++ packages/data-table/src/lib/type-safety.test.ts
@@ -18,7 +18,8 @@ import type { CountQuery as _CountQuery } from '../index.ts'
 import type { InsertCommand as _InsertCommand } from '../index.ts'
 // @ts-expect-error QueryMethod is no longer exported
 import type { QueryMethod as _QueryMethod } from '../index.ts'
-import { table, hasMany } from './table.ts'
+import { table } from './table.ts'
+import { hasMany } from './table-relations.ts'
 import type { TableReference, TableRow } from './table.ts'
 import { eq } from './operators.ts'
 import type { SqliteTestSeed } from '../../test/sqlite-test-database.ts'
